Add exception handling for Electron
There are some edge cases where `Object.defineProperty` throws an error.

For example Electron's native elements such as `webview` throws an error: "Uncaught TypeError: Cannot redefine property: src" when cell tries to `Nucleus.set` via `Object.defineProperty`.

In most cases this is not problematic since this never happens with regular browsers. Not sure what the root cause for this is on electron but for now just wrapping it inside try/catch to get around this. Seems to work fine this way.